Adds a has submission icon to Predictor
This addresses a bug where we were not summing predicted points for any closed assignments,
even if we had enabled the slider. This happened because we were not passing information
about the student's submissions to the front end.

This adds a file icon in the predictor icon bar if the student has made a submission for
the assignment, and only discounts predicted points for closed assignments without
a student submission.

Additionally, this updates the sample data to more explicitly test all areas of concern
in the predictor, and to better describe what is being tested within the assignment title
and description. It also fixes a number of bugs which prohibited predicted data from
accurately representing model states found in the real world.