bugfix for group rubric grades missing submission id
This addresses the bug where rubric grades for groups did not persist the submission id.
This is because the service was looking for a submission with the student id and returning
nil. This adds the group to the service context, and checks to see if a group is available.
If it is, it will check for a group submission, otherwise it will default to the previous
behavior.

closes #1752