Fix Broken Rspec Tests
The default submission factory assumes that it is for an individual,
meaning it will have a student_id and not a group_id.

Since a valid Submission is defined as having either a student_id or a
group_id, instances where we were testing group submissions needed
to reference a new group_submission factory where student_id = nil.