Return 401 vs redirecting if API request
Per the Angular docs regarding $http usage:

Note that if the response is a redirect, XMLHttpRequest will
transparently follow it, meaning that the outcome (success or error)
will be determined by the final response status code.

If the user is unauthorized due to an OAuth issue, a 401 is returned for
handling by the client.