🎉🎉 Fixed the bug?????????
After upgrading to Rails 5.2, the history-related spec tests began to
fail. It was determined that paper_trail was being loaded in its current
implementation twice on some models, thereby causing duplicate versions
to be created.

This commit revises the ActiveRecord::Concern so that has_paper_trail
must be added to the model instead of automatically being included
in the Historical module, so that if any options are needed they can
be set there.